requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
scanpy>=1.8.0
torch>=1.10.0
torch-geometric>=2.0.0
networkx>=2.6.0
umap-learn>=0.5.0
adjustText  # Optional for better label placement